requirements.txt
fastapi
uvicorn
python-multipart
faiss-cpu
sentence-transformers
pymupdf
langchain
langchain-community
tiktoken
nltk
openai
numpy
pandas
pydantic
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
databases
python-dotenv
loguru
httpx
